refactor(ReservationService): 추상화 삭제
