+++ src/main/java/roomescape/service/ReservationService.java
@@ -1,15 +1,47 @@
 package roomescape.service;
 
 import java.util.List;
+import org.springframework.stereotype.Service;
+import roomescape.dao.reservation.ReservationDao;
+import roomescape.dao.resetvationTime.ReservationTimeDao;
+import roomescape.domain.ReservationTime;
 import roomescape.dto.request.ReservationCreateRequest;
 import roomescape.dto.response.ReservationCreateResponse;
 import roomescape.dto.response.ReservationResponse;
+import roomescape.dto.response.ReservationTimeResponse;
 
-public interface ReservationService {
+@Service
+public class ReservationService {
 
-    ReservationCreateResponse create(ReservationCreateRequest reservationCreateRequest);
+    private final ReservationDao reservationDao;
+    private final ReservationTimeDao reservationTimeDao;
 
-    List<ReservationResponse> findAll();
+    public ReservationService(final ReservationDao reservationDao, final ReservationTimeDao reservationTimeDao) {
+        this.reservationDao = reservationDao;
+        this.reservationTimeDao = reservationTimeDao;
+    }
 
-    void delete(final Long id);
+    public ReservationCreateResponse create(ReservationCreateRequest reservationCreateRequest) {
+        ReservationTime time = reservationTimeDao.findById(reservationCreateRequest.timeId());
+        return new ReservationCreateResponse(
+                reservationDao.create(reservationCreateRequest),
+                reservationCreateRequest.name(),
+                reservationCreateRequest.date(),
+                new ReservationTimeResponse(time.getId(), time.getStartAt()));
+    }
+
+    public List<ReservationResponse> findAll() {
+        return reservationDao.findAll().stream()
+                .map(reservation -> new ReservationResponse(
+                        reservation.getId(),
+                        reservation.getName(),
+                        reservation.getDate(),
+                        reservation.getTime()
+                ))
+                .toList();
+    }
+
+    public void delete(final Long id) {
+        reservationDao.delete(id);
+    }
 }
